refactor: move block requirements logic into framework
