+++ gateway-framework/src/gateway/http/requests/mod.rs
@@ -1,4 +1,5 @@
 pub mod auth;
+pub mod blocks;
 mod selector;
 
 pub use selector::{RequestSelector, RequestSelectorRejection};
